backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
langchain-core
langchain-anthropic
langchain-community
langgraph
ddgs
